+++ routes/adminStrategy.js
@@ -63,7 +63,6 @@ router.post('/addStrategy', function (req, res, next) {
         strategy.hasUserAndGame(data, function (result) {
             if (result.admin) {
                 data.game_name = data.game_name || null
-                // data.add_time = date.Format("yyyy-MM-dd HH:mm:ss") || null
                 data.add_time = parseInt(date.getTime() / 1000);
                 data.adminstatus = 1;
                 data.img_src = "http://img.oneyouxi.com.cn/" + data.img_src;
@@ -170,12 +169,14 @@ router.get('/deleteStrategy', function (req, res) {
             var imgReg = /<img.*?(?:>|\/>)/gi;
             var srcReg = /src=[\'\"]?([^\'\"]*)[\'\"]?/i;
             var arr = str.match(imgReg);
-            for (var i = 0; i < arr.length; i++) {
-                var src = arr[i].match(srcReg);
-                //获取图片地址，src[1]
-                var imgUrl = src[1].substr(src[1].indexOf("www"))//获取图片地址，并从www开始截取后面的字符
-                if (fs.existsSync(path.join(__dirname, '../' + imgUrl))) {//查看文件是否存在，是返回true，否返回fales
-                    fs.unlinkSync(path.join(__dirname, '../' + imgUrl));//执行删除文件
+            if (arr) {
+                for (var i = 0; i < arr.length; i++) {
+                    var src = arr[i].match(srcReg);
+                    //获取图片地址，src[1]
+                    var imgUrl = src[1].substr(src[1].indexOf("www"))//获取图片地址，并从www开始截取后面的字符
+                    if (fs.existsSync(path.join(__dirname, '../' + imgUrl))) {//查看文件是否存在，是返回true，否返回fales
+                        fs.unlinkSync(path.join(__dirname, '../' + imgUrl));//执行删除文件
+                    }
                 }
             }
